+++ src/components/InstallShell.tsx
@@ -1,6 +1,7 @@
 import type { ReactNode } from 'react'
 import { InstallPromptProvider } from '../context/InstallPromptProvider'
 import { InstallInviteSheet } from './install/InstallInviteSheet'
+import { InstallFloatingPill } from './install/InstallFloatingPill'
 import { useTracker } from '../context/trackerContext'
 import { useToast } from '../hooks/useToast'
 
@@ -17,6 +18,7 @@ export function InstallShell({ children }: { children: ReactNode }) {
       }
     >
       {children}
+      <InstallFloatingPill />
       <InstallInviteSheet />
     </InstallPromptProvider>
   )